Moved symbols.hpp
