+++ include/calculate/symbols.hpp
